requirements.txt
# This file was autogenerated by uv via the following command:
#    ./scripts/generate_requirements.sh
agno==2.4.7
agno-infra==1.0.7
annotated-doc==0.0.4
annotated-types==0.7.0
anthropic==0.77.1
anyio==4.12.1
attrs==25.4.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cryptography==46.0.4
distro==1.9.0
dnspython==2.8.0
docstring-parser==0.17.0
duckdb==1.4.4
duckduckgo-search==8.1.1
email-validator==2.3.0
fastapi==0.128.0
fastapi-cli==0.0.20
fastapi-cloud-cli==0.11.0
fastar==0.8.0
gitdb==4.0.12
gitpython==3.1.46
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
httpx-sse==0.4.3
hyperframe==6.1.0
idna==3.11
importlib-metadata==8.7.1
jinja2==3.1.6
jiter==0.12.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
lxml==6.0.2
markdown-it-py==4.0.0
markupsafe==3.0.3
mcp==1.26.0
mdurl==0.1.2
numpy==2.4.1
openai==2.16.0
openinference-instrumentation==0.1.43
openinference-instrumentation-agno==0.1.27
openinference-semantic-conventions==0.1.25
opentelemetry-api==1.39.1
opentelemetry-instrumentation==0.60b1
opentelemetry-sdk==1.39.1
opentelemetry-semantic-conventions==0.60b1
packaging==26.0
pgvector==0.4.2
pillow==12.1.0
primp==0.15.0
psycopg==3.3.2
psycopg-binary==3.3.2
pycparser==3.0
pydantic==2.12.5
pydantic-core==2.41.5
pydantic-extra-types==2.11.0
pydantic-settings==2.12.0
pygments==2.19.2
pyjwt==2.10.1
python-dotenv==1.2.1
python-multipart==0.0.22
pyyaml==6.0.3
referencing==0.37.0
reportlab==4.4.9
rich==14.3.1
rich-toolkit==0.17.1
rignore==0.7.6
rpds-py==0.30.0
sentry-sdk==2.51.0
shellingham==1.5.4
smmap==5.0.2
sniffio==1.3.1
sqlalchemy==2.0.46
sse-starlette==3.2.0
starlette==0.50.0
tomli==2.4.0
tqdm==4.67.1
typer==0.21.1
typing-extensions==4.15.0
typing-inspection==0.4.2
urllib3==2.6.3
uvicorn==0.40.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==16.0
wrapt==1.17.3
zipp==3.23.0
